Scroll managing overhaul.
Scroll managers containing the modifiers must be more easily
manipulated so that the controls aren't as difficult to use.

Content inside scroll managers is now able to show any kind
of content, which is kept when selecting it and hidden when
not hovering over it.

Scroll managers containing the bug's modifiers now animate
correctly.