Adds level completion screens and cine machine camera.
Levels need to have a final screen that shows the number of
stars won by the player, and allow him to get to the next
level.

Adds and replaces some shapes for the interface.

Camera is now managed by cine machine instead of the
custom follow camera behaviour. This change requires the
input manager to raycast only certain objects when
detecting modifiable entities.

Damage dealers now teleport the player when it takes damage.

A new scene test is created.