Corrects modifier placement and adds bug correction
When the camera is moved, modifier removal from bugs stops
working, this shouldn't be happening.

When the player is damaged, the bugs should be corrected in
order, from left to right, deleting the modifiers the player
added from top to bottom.

Bugs are now deleted when the player takes damage, and the
modifier removal now works properly.

MISSING : When clicking a modifier inside a bug with LMB, it
bugs out. Bug removal on logical isn't automated, and logic
agents must correct it themselves.