Level 1 creation.
An introductory level, showing the mechanics, basic controls
and puzzles must be created.

Creates the level 1, level in which the player is introduced
to the mechanics and is presented a basic puzzle.

Level 1 must not have the correction mechanic, so a new flag
is added to the Bug Manager so that it can control if the
level corrects the bugs or not.