Level 2 creation.
Level 2 must be created, along with the needed objects to
provide a challenge to the player.

Level 2 combines the speed and magnetism modifiers, along
with bugable buttons and levers, to provide a not so obvious
challenge.

Creates the logic for buttons, levers and activable platforms,
plus hanging spike balls, bugable or not.