Adds audio attributes and attack capabilities.
Audio clip attributes need to be added to most of the entities
and objects, so that SFX can be correctly played.

Damageable now contains audio attributes for attacking, taking
damage and dying. Main Character script now plays death audio.

Main Character controller os updated to correctly set the
moving booleans.

Lets the entities now decide if they are facing right or left
from the editor, so that enemies that can look both ways can
be created.

Main Character script now has all the attacker attributes,
and the enemy now has a push effect.

Corrects a bug in the doors, which would launch an exception
when not finding the bug.

Level manager is updated so it can freeze entities in place
when starting the level.

UI tracker bug where it issued exceptions when not finding
a target corrected.

Object scripts now add an audio clip that is played when
using them.